pyproject.toml
[project]
name = "awesome-ai-news"
version = "1.0.0"
description = "Automated AI news aggregator with intelligent clustering and curation"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "CC0-1.0" }
authors = [
    { name = "awesome-ai-news-bot" }
]
dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "google-generativeai>=0.8.0",
    "feedparser>=6.0.0",
    "aiohttp>=3.11.0",
    "pyyaml>=6.0",
    "python-slugify>=8.0.0",
    "tenacity>=9.0.0",
    "loguru>=0.7.0",
    "typer>=0.15.0",
    "python-dotenv>=1.2.1",
    "google-genai>=1.56.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-bdd>=8.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-bdd>=8.0.0",
    "mypy>=1.13.0",
    "pyright>=1.1.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "aioresponses>=0.7.6",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["**/__pycache__", ".venv", "build", "dist"]
reportMissingImports = true
reportMissingTypeStubs = false

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "bdd: BDD tests",
    "slow: Slow running tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
